Move classes de teste para o pacote correto
