+++ src/br/com/caelum/financas/modelo/teste/TestaEdicaoConta.java
@@ -1,25 +1,26 @@
-package br.com.caelum.financas.modelo;
+package br.com.caelum.financas.modelo.teste;
 
 import javax.persistence.EntityManager;
 
 import br.com.caelum.financas.dao.ContaDao;
 import br.com.caelum.financas.infra.JPAUtil;
+import br.com.caelum.financas.modelo.Conta;
 
-public class TestaBuscaConta {
+public class TestaEdicaoConta {
 
 	public static void main(String[] args) {
-		EntityManager manager = new JPAUtil().getEntityManager();
-		
-		ContaDao dao = new ContaDao(manager);
-		
 		Conta novaConta = new Conta("Alexandre Gama", "1234567", "123");
+		
+		EntityManager manager = new JPAUtil().getEntityManager();
+		ContaDao dao = new ContaDao(manager);
 		manager.getTransaction().begin();
 		dao.adiciona(novaConta);
-		manager.getTransaction().commit();
 		
 		Conta conta = dao.buscaPorId(novaConta.getId());
+		conta.setTitular("Gama");
+		manager.merge(conta);
 		
-		System.out.println(conta);
+		manager.getTransaction().commit();
 		manager.close();
 	}
